cloud_tests: help pylint
pylint missed finding a typo in the lxd platform because it could not
determine that the variable was being used was a string.  The variable
was set by loading a yaml file which pylint couldn't know that it
would be a string.  In these cases, we can be more explicit.