subp: support combine_capture argument.
This adds 'combine_capture' argument as was present in curtin's
subp.  It is useful to get interleaved output of a command.  I noticed
a need for it when looking at user_data_rhevm in DataSourceAltCloud.
That will run a subcommand, logging its stdout but swallowing its stderr.

Another thing to change to use this would be in udevadm_settle which
currently just returns the subp() call.

Also, add the docstring copied from curtin's subp.