tests: drop salt-minion integration test
The salt minion integration test as we had it did not do a whole lot
more than the unit tests on that module did.  Additionally, it caused
some transient failures at least in Ubuntu 18.04.

At a future date we may choose to add an integration test that installs
salt-minion and salt server and configures it to be a better test.

LP: #1778737