tests: disable snap, fix write_files binary
The snap test requires access to a proxy and currently the integration
tests do not handle this scenario. I am disabling the test untill I can
loop back around and fix this.

The write_files test, specifically, the binary test is failing on cosmic
because the "binary" file we were writting was not a complete elf
executable, but we expected 'file' to identify it as such.

The change here is to simply use some 24 bytes of random, non-utf data
and check that file was written correctly via expected checksum.