hotplug: fix socket/service units and logging
Rework socket/service units and dependencies.  The socket
will be active and the service only depends on the socket.

The service file uses bash to read args from fd=3 and exec
cloud-init with the arguments.  The service exec exits 0 to
prevent hotplug errors from shutting down the socket.

Replace hotplug-hook --debug arg (which is part of main) with
--hotplug-debug to prevent cloud-init from configuring logging
more than once.

Use a logging wrapper to emit hotplug-hook messages to stderr
and to the configured log.  The allows the hotplugd service to
emit messages and directs the verbose cloud-init output into
/var/log/cloud-init.log