Hotplug: restrict to pci devices, query datasource for update events
Narrow the types of devices that cloud-init will attempt to handle.

Replace the use of a file marker for enabling hotplug handling, instead
have the datasource example system and user config for precense of
'updates' config which indicates which subsystems and which event types
the system/user wants cloud-init to handle.