config: On ubuntu select cloud archive mirrors for armel, armhf, arm64.
Infrastructure is now set up for Ubuntu to handle Amazon instances
hitting a ports archive at:
   - http://%(ec2_region)s.ec2.ports.ubuntu.com/ubuntu-ports/

And additionally, generic mirrors at
   *.clouds.ports.ubuntu.com/ubuntu-ports

The change here will utilize those mirrors for the arm64, armel and
armhf arches.

We've decided to limit the auto-selection of those
mirrors to arm, where we know a use case.  That way new instances
of ppc64el or other arches will not select them. Such a behavior change
could be problematic for a user in a firewalled environment.

LP: #1805854