lxd: install zfs-linux instead of zfs meta package
When using the LXD module cloud-init will attempt
to install ZFS if it does not exist on the target
system. However instead of installing the `zfsutils-linux`
package it attempts to install `zfs` resulting in an error.

Ubuntu Xenial (16.04) has zfs meta package, but Bionic (18.04)
does not. Use the specific base package instead of zfs meta.

Co-authored-by: Michael Skalka <michael.skalka@canonical.com>

LP: #1799779