tools/read-version: handle errors
When the cloned branch was not the canonical upstream and tags were not
available, tox would fail because tools/read-version would fail, and
tragically never print the advice that is in tools/read-version about
how to fix it.

This changes tools/read-version to catch the exception that is elsewhere
explicitly thrown and treat that too as an error it can handle.