network: Fix type and respect name when rendering vlan in sysconfig.
Prior to this change, vlans were rendered in sysconfig with
'TYPE=Ethernet', and incorrectly renders the PHYSDEV based on
the name of the vlan device rather than the 'link' provided
in the network config.

The change here fixes:
 * rendering of TYPE=Ethernet for a vlan
 * adds a warning if the configured device name is not supported
   per the RHEL 7 docs "11.5. Naming Scheme for VLAN Interfaces"

Forward-ported to cloud-init-20.2: Minimal changes on the order of
things on cloudinit/net/sysconfig.py.

LP: #1788915
LP: #1826608
rhbz: #1861871

Signed-off-by: Scott Moser <smoser@brickies.net>
Signed-off-by: Eduardo Otubo <otubo@redhat.com>