Add a cdrom size checker for OVF ds to ds-identify
With a large size ISO file attached to iso dev, ds-identify might
grep it entirely if iso dev is ISO9660, it takes very long time to
start OS.

Resolve this by:

- Adding a checker to read the ISO size (from sysfs).  If
the size of the ISO filesystem is > 10MiB then the ISO will be
ignored (logged as oversized).

- Move the ovf vmware guest customization checker to be
ahead of cdrom ovf checker, so no need check the ISO size if vmware
guest customization is enabled.

LP: #1806701