templates/ntp.conf.debian.tmpl: fix missing newline for pools
The debian ntp.conf template did not contain a newline for the
comment used to mark the rendered ntp pools configured.  This
resulted in an invalid line:

'# poolspool 0.int.pool.ntp.org iburst'

rather than:

'# pools
 pool 0.int.pool.ntp.org iburst'

This patch fixes the template and updates the unittest to
verify that the rendered templates puts servers and pools
at the beginning of a line.

LP: #1836598