debian/cloud-init.config: fix yaml parsing with whitespace
Users may rewrite valid yaml adding or omitting additional white space
when defining the 'datasources_list' variable. If the yaml did not
include a space between the start of the list '[' and the members and
the end ']' then this would break the shell yaml parsing and produce
invalid yaml on dpkg-reconfigure cloud-init. Fix this issue by updating
the sed matching expression to allow for zero or more spaces after the
leading bracket and before the trailing bracket. For example, this
branch now accepts:

  datasources_list: [NoCloud, None]

LP: #1841697