Brightbox: restrict detection to require full domain match .brightbox.com
The detection for brightbox in both ds-identify and in
identify_brightbox would incorrectly match the domain 'bobrightbox',
which is not a brightbox platform.  The fix here is to restrict
matching to '*.brightbox.com' rather than '*brightbox.com'

Also, while here remove a url to bug 1661693 which added the
knowledge of brightbox.