net: add is_master check for filtering device list
Some network devices are transformed into a bond via kernel magic
and do not have the 'bonding' sysfs attribute, but like a bond they
have a duplicate MAC of other bond members.  On Azure Advanced
Networking SRIOV devices are auto bonded and will have the same MAC
as the HyperV nic.  We can detect this via the 'master' sysfs attribute
in the device sysfs path and this patch adds this to the list of devices
we ignore when enumerating device lists.

LP: #1844191