get_interfaces: don't exclude bridge and bond members
The change that introduced this issue was handling interfaces that are
bonded in the kernel, in a way that doesn't present as "a bond" to
userspace in the normal way.  Both members of this "bond" will share a
MAC address, so we filter one of them out to avoid incorrect MAC address
collision warnings.

Unfortunately, the matching condition was too broad, so that change also
affected normal bonds and bridges.  This change specifically excludes
bonds and bridges from that determination, to address that regression.

LP: #1846535