debian/cloud-init.config get_yaml_list should return only the value
When fixing LP: #1841697 we mistakely added the key and list brackets
to the return value of the get_yaml_list function.  This resulted in
cloud-init.postinst which creates 90_dpkg.cfg to duplicate the key
and brackets resulting in an invalid yaml file.  This commit removes
the emitted $key and [ ] portions resulting in valid yaml when
cloud-init.postinst renders 90_dpkg.cfg.

LP: #1846511