exoscale: Increase url_max_wait to 120s.
The exoscale datasource defines a shorter timeout than the default (10)
but did not override url_max_wait, resulting in a single attempt being
made to wait for the metadata service.

In some rare cases, a race condition means the route to the metadata
service is not set within 10 seconds, and more attempts should be made.

This sets the url_max_wait for the datasource to 120.