FreeBSD fix for get_linux_distro()
a lot of our code assumes that we're on Linux, and all it needs to know
then is which exact distro it's on. We have structured our code such,
that FreeBSD *is* a distro, so before resorting to return nothing on
FreeBSD whenever get_linux_distro() is queried, we check if we're on
FreeBSD, and return distro_name = 'freebsd' and 'distro_version' with
the result of `freebsd-version`.

lp: #1815030