add unit test for get_linux_distro() → FreeBSD
we add a unit test to verify if our "Linux Distro" is FreeBSD
To do this, we need to ensure that the file '/bin/freebsd-version'
"exists", and that `uname -r` returns something sensible.

in adding this test, i also "discovered" that `subp` returns stdout and
stderr, and i need to split them, and strip() stdout.

For now, optimistically, no check is made if stderr != empty.