util: read_seeded drop unused and incorrent positional file_retries
read_file_or_url does not accept a parameter named file_retries.
If a path is provided as base, then read_file_or_url ignores retries
parameter and only attempts to read the file once.

Drop file_retries as it is unused, and was incorrectly being passed as
a positional parameter that mapped to headers, but was thankfully an
ignored parameter as well in the case of file paths.

read_seeded is used by NoCloud, CloudStack and OVF datasources, all of
which provide file paths instead of URLs, so this invalid file_retries
parameter hadn't gotten used.