Do not use fallocate in swap file creation on xfs.
When creating a swap file on an xfs filesystem, fallocate cannot be used.
Doing so results in failure of swapon and a message like:
 swapon: swapfile has holes

The solution here is to maintain a list (currently containing only XFS)
of filesystems where fallocate cannot be used. The, on those fileystems
use the slower but functional 'dd' method.

Based on initial pull request:
https://code.launchpad.net/~adobrawy/cloud-init/+git/cloud-init/+merge/354680

And further comments from Scott Moser:
http://paste.ubuntu.com/p/8pVhczVqG3/

Signed-off-by: Eduardo Otubo <otubo@redhat.com>