Support Centos dhclient lease format for option 121
This is linked to the bug 1850642
https://bugs.launchpad.net/cloud-init/+bug/1850642

Basically, redhat flavor images use a different name for the
classless-static-routes option (121) which isn't the same one as for
Debian.  This cause the conflict with the dhclient-script used as well.

With this, the option would be supported on both Redhat flavored and
debian flavored OS (with a preference for Debian as being the first in
the list).

Reference to lookup as well;
https://bugzilla.redhat.com/show_bug.cgi?id=516325
https://github.com/vaijab/fedora-dhcp/blob/e83fb19c51765442d77fa60596bfdb2b3b9fbe2e/dhcp-rfc3442-classless-static-routes.patch#L252
https://github.com/heftig/NetworkManager/blob/f56c82d86122fc45304fc829b5f1e4766ed51589/src/dhcp-manager/nm-dhcp-client.c#L978