docs: fix line length and remove highlighting
doc8 does not know about the ephasize-lines portion of code-block and
throws an error. As this is the only place right now I am going to
remove it untill we can find a better solution. rstcheck and
restructuredtext-lint have issues with sphinx declaritives, so
doc8 is still the best to use for now.