Remove top-level random import to allow VM to boot faster
As the machine starts, the VM doesn't have the /dev/urandom initialised
yet.  This cause the init-local dhcp discovery to wait on it to reach
the meta-data server and start configuring the server all to resolve
169.254.169.254 which shouldn't need resolution.  The impact of this is
80 sec on Redhat based OS.

As python import are cached anyway and random is only used in "rand_str"
, putting the import within the function that needs it makes sense at
the moment of writing.

May come a time where we might want to have a test to make sure random
and other packages having similar effect never gets imported as a
top-level package in this file being used throughout cloud-init to avoid
this problem again.