Refactoring create_swapfile and adding create_swap
In order to avoid duplicate code the function create_swapfile() was
reduced and a new function create_swap() was introduced.

create_swapfile identifies which type of filesystem is currently in use
(xfs or btrfs) and calls create_swap(), that chooses between dd and
fallocate - this, in case of failure will try dd as fallback.

Signed-off-by: Eduardo Otubo <otubo@redhat.com>