Only use gpart if it is the BSD gpart
Currently, cloud-init will happily try to run `gpart` on Linux even
though on most distributions this a different tool [1]. Extend the
availability check to make sure the `gpart` present is really the BSD
variant, to avoid accidental execution.

Also add a pointer to the docs, so that people do not try to install
gpart on Linux in the expectation it will work with this module.

[1] https://github.com/baruch/gpart